app(redux): Rename reducers
