+++ app/redux/reducers/cacheReducer.js
