+++ app/redux/reducers/cinemaReducer.js
