+++ app/redux/reducers/watcherReducer.js
